Data.Tuple.Append.Class

Description

A module that defines typeclasses to prepend and append items and tuples into new tuples.

Add an element to a tuple

class TupleAddL x 𝐯 x𝐯 | x 𝐯 -> x𝐯, x𝐯 -> x, x𝐯 -> 𝐯 where

A typeclass mainly used to construct a tuple with one element extra. That element is added at the left side of the tuple. The typeclass is also used for a small amount of extra datatypes to make it more convenient.

Methods

(<++) infixr 5

Arguments

:: x
The item to prepend at the left side of the tuple.

-> 𝐯
The tuple containing the rest of the elements.

-> x𝐯
A tuple that has one element more than the given tuple: the given item that is prepended at the left side.

Construct a new tuple by adding the first parameter as first item in the tuple.

class TupleAddR 𝐯 x 𝐯x | 𝐯 x -> 𝐯x, 𝐯x -> 𝐯, 𝐯x -> x where

A typeclass mainly used to construct a tuple with one element extra. That element is added at the right side of the tuple. The typeclass is also used for a small amount of extra data types to make it more convenient.

Methods

(++>) infixl 5

Arguments

:: 𝐯
The tuple containing the rest of the elements.

-> x
The item to append at the right side of the tuple.

-> 𝐯x
A tuple that has one element more than the given tuple: the given item that is appended at the right side.

Construct a new tuple by adding the second parameter as last item in the tuple.

Append two tuples

class TupleAppend 𝐮 𝐯 𝐮𝐯 | 𝐮 𝐯 -> 𝐮𝐯, 𝐮 𝐮𝐯 -> 𝐯, 𝐯 𝐮𝐯 -> 𝐮 where

A typeclass mainly used to append two tuples together into a tuple that contains as many elements as the sum of the number of elements of the two given tuples. The typeclass is also used for a small amount of extra data types to make it more convenient.

Methods

(+++) infixr 5

Arguments

:: 𝐮
The first tuple to append.

-> 𝐯
The second tuple to append.

-> 𝐮𝐯
A tuple that contains the items of the first and the second tuple.

Construct a new tuple that contains the elements of the two given tuples.

Lifting tuples of applicatives

class Applicative f => SequenceTuple f f𝐮 𝐮 | f𝐮 -> f 𝐮, f f𝐮 -> 𝐮, f 𝐮 -> f𝐮 where

A typeclass to process a tuple of Applicative elements to an Applicative of a tuple. While a 2-tuple has a sequenceA function, that function sees the tuples as a collection of one element: the second item. This SequenceTuple typeclass considers this a collection of n elements for an n-tuple and thus runs over all elements of the tuple.

Minimal complete definition

sequenceTupleA

Methods

sequenceTupleA

Arguments

:: f𝐮
The tuple with applicative elements.

-> f 𝐮
An applicative tuple thas has sequenced over the elements of the tuple.

Sequence the elements of the tuple. For an n tuple sequenceTupleA (v₁, v₂, …, vₙ) is equivalent to: (,,…,) $ v₁ * v₂ * … * vₙ.

default sequenceTupleA :: (Traversable t, 𝐮 ~ t b, f𝐮 ~ t (f b)) => f𝐮 -> f 𝐮

sequenceTupleA_

Arguments

:: f𝐮
The tuple of applicatives to sequence.

-> f ()
An applicative for the unit type.

Sequence the elements of the tuple, and return the unit. For an n tuple sequenceTupleA_ (v₁, v₂, …, vₙ) is equivalent to: v₁ *> (v₂ *> (… *> (vₙ *> pure ()))).

Folding elements in a tuple

class FoldTuple v 𝐯 | 𝐯 -> v where

A typeclass to fold a tuple of an arbitrary length. It thus implements the equivalent of foldl, foldr and foldMap on tuples, like it would have done on a list with an equivalent length. This can only be done if the type of all the elements of the tuple is the same. The v type variable is type of the items of the tuple, whereas 𝐯 is the type of the tuple.

Minimal complete definition

foldMapTuple | foldrTuple

Methods

foldlTuple

Arguments

:: (a -> v -> a)
The "folding function" that takes the acculator thus far and an element from the tuple, and produces a new accumulator.

-> a
The initial value for the accumulator to use.

-> 𝐯
The tuple that we "fold".

-> a
The result of the folding process.

Fold any tuple left-to-right with the given folding function that folds a second element, the first value for the accumulator and the tuple to fold, so:

foldlTuple f z (v₁, v₂, …, vₙ) == f (… f (f (f z v₁) v₂) …) vₙ

foldrTuple

Arguments

:: (v -> a -> a)
The "folding function" that takes an element from the tuple, the accumulator, and produces a new accumulator.

-> a
The initial value for the accumulator to use.

-> 𝐯
The tuple that we "fold".

-> a
The result of the folding process.

Fold any tuple right-to-left with the given folding function that folds a second element, the first value for the accumulator and the tuple to fold, so:

foldrTuple f z (v₁, v₂, …, vₙ) == f v₁ (f v₂ (… (f vₙ z) …)))

foldMapTuple

Arguments

:: Monoid m

=> (v -> m)
The mapping function that maps the elements of the tuple to a value of a Monoid type.

-> 𝐯
The tuple to "fold".

-> m
The result of the folding process.

Map the items in the tuple to a value of a Monoid type and then fold these through the Monoid instance.

foldMapTuple f z (v₁, v₂, …, vₙ) == f v₁ <> (f v₂ <> (… (f vₙ <> mempty) …)))