Data.Tuple.Append.Class

Description

A module that defines typeclasses to prepend and append items and tuples into new tuples.

Add an element to a tuple

class TupleAddL x 𝐯 x𝐯 | x 𝐯 -> x𝐯, x𝐯 -> x, x𝐯 -> 𝐯 where

A typeclass mainly used to construct a tuple with one element extra. That element is added at the left side of the tuple. The typeclass is also used for a small amount of extra datatypes to make it more convenient.

Methods

(<++) infixr 5

Arguments

:: x
The item to prepend at the left side of the tuple.

-> 𝐯
The tuple containing the rest of the elements.

-> x𝐯
A tuple that has one element more than the given tuple: the given item that is prepended at the left side.

Construct a new tuple by adding the first parameter as first item in the tuple.

class TupleAddR 𝐯 x 𝐯x | 𝐯 x -> 𝐯x, 𝐯x -> 𝐯, 𝐯x -> x where

A typeclass mainly used to construct a tuple with one element extra. That element is added at the right side of the tuple. The typeclass is also used for a small amount of extra data types to make it more convenient.

Methods

(++>) infixl 5

Arguments

:: 𝐯
The tuple containing the rest of the elements.

-> x
The item to append at the right side of the tuple.

-> 𝐯x
A tuple that has one element more than the given tuple: the given item that is appended at the right side.

Construct a new tuple by adding the second parameter as last item in the tuple.

Append two tuples

class TupleAppend 𝐮 𝐯 𝐮𝐯 | 𝐮 𝐯 -> 𝐮𝐯, 𝐮 𝐮𝐯 -> 𝐯, 𝐯 𝐮𝐯 -> 𝐮 where

A typeclass mainly used to append two tuples together into a tuple that contains as many elements as the sum of the number of elements of the two given tuples. The typeclass is also used for a small amount of extra data types to make it more convenient.

Methods

(+++) infixr 5

Arguments

:: 𝐮
The first tuple to append.

-> 𝐯
The second tuple to append.

-> 𝐮𝐯
A tuple that contains the items of the first and the second tuple.

Construct a new tuple that contains the elements of the two given tuples.